Revert "feat(blazorui): add immediate functionalities to BitOtpInput #8566 (#9943)"
This reverts commit ee0824037f9a3b226df5d729a58125a457dcbe59.
