+++ src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Inputs/OtpInput/BitOtpInputDemo.razor.scss
@@ -1,11 +1,4 @@
-﻿@import '../../../../Styles/abstracts/_functions.scss';
-@import '../../../../Styles/abstracts/_bit-css-variables.scss';
-
-.example-box {
-    display: flex;
-    max-width: rem2(300px);
-    flex-direction: column;
-}
+﻿@import '../../../../Styles/abstracts/_bit-css-variables.scss';
 
 ::deep {
     .custom-class {
